db: moved db/store to infrastructure folder
