+++ cmd/ports-service/handlers/handlers.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/alexlucaci-go/ports-service/domain/ports"
-	"github.com/alexlucaci-go/ports-service/domain/ports/store/inmemorydb"
+	"github.com/alexlucaci-go/ports-service/infrastructure/store/inmemorydb"
 	"github.com/alexlucaci-go/ports-service/web"
 )
