+++ cmd/ports-service/main.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/alexlucaci-go/ports-service/cmd/ports-service/handlers"
 	"github.com/alexlucaci-go/ports-service/domain/ports"
-	"github.com/alexlucaci-go/ports-service/domain/ports/store/inmemorydb"
+	"github.com/alexlucaci-go/ports-service/infrastructure/store/inmemorydb"
 	"github.com/alexlucaci-go/ports-service/loader"
 )
