+++ cmd/ports-service/tests/ports_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/alexlucaci-go/ports-service/cmd/ports-service/handlers"
 	"github.com/alexlucaci-go/ports-service/domain/ports"
-	"github.com/alexlucaci-go/ports-service/domain/ports/store/inmemorydb"
+	"github.com/alexlucaci-go/ports-service/infrastructure/store/inmemorydb"
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/require"
 )
